PuLP (api) solvers comparison: CBC, GLPK, HiGHS
In conclusion, the HiGHS solver seems like the fastest general solver. However, the other solvers can also be fast (or faster) if the integer constraints are relaxed. This shows that the choice of solver and its settings can have a significant impact on the solution time and the quality of the solution. It is important to experiment with different solvers and settings to find the best combination for a given problem.